+++ src/EnrolmentWorkflow/Rule/Factory.php
@@ -15,6 +15,7 @@ use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\NoResponsibleUsersAssign
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\OrgUnitSuperior\OrgUnitSuperior;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\OrgUnitUserType\OrgUnitUserType;
 use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\UDF\UDF;
+use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\UDFSupervisor\UDFSupervisor;
 use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
 
 /**
@@ -64,7 +65,8 @@ final class Factory
             NoResponsibleUsersAssigned::class,
             OrgUnitSuperior::class,
             OrgUnitUserType::class,
-            UDF::class
+            UDF::class,
+            UDFSupervisor::class
         ];