+++ src/EnrolmentWorkflow/Rule/UDFSupervisor/UDFSupervisor.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\UDFSupervisor;
+
+use srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\UDF\UDF;
+
+/**
+ * Class UDFSupervisor
+ *
+ * @package srag\Plugins\SrUserEnrolment\EnrolmentWorkflow\Rule\UDFSupervisor
+ *
+ * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
+ */
+class UDFSupervisor extends UDF
+{
+
+    const TABLE_NAME_SUFFIX = "udfs";
+}